pyproject.toml
[tool.ruff]
line-length = 88
fix = true

[tool.ruff.lint]
select = ["E", "F", "I"]  # "E" = pycodestyle, "F" = pyflakes, "I" = isort

[tool.setuptools]
packages = []

[project]
name = "ExecutiveLLM"
version = "1.0.0"
description = "ExecutiveLLM - Benchmarks for LLMs in management and decision making tasks"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.39.0",
    "openai>=1.0.0",
    "google-generativeai>=0.8.0",
    "python-dotenv>=1.0.0",
    "axelrod>=4.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ty==0.0.1a27",
]

[tool.ty.environment]
root = ["benchmarks/axl-bench"]

[tool.ty.src]
include = ["src/**/*.py"]
exclude = []
